+++ src/main/resources/mapper/company.xml
@@ -3,7 +3,7 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="site.metacoding.miniproject.domain.company.CompanyDao">
 
-	<insert id="insert" parameterType="site.metacoding.miniproject.domain.company.Company" useGeneratedKeys="true" keyProperty="companyId">
+	<!-- <insert id="insert" parameterType="site.metacoding.miniproject.domain.company.Company" useGeneratedKeys="true" keyProperty="companyId">
 		INSERT INTO company(company_number, company_name, company_email, company_tel, company_location, company_username, company_password)
 		VALUES(#{companyNumber}, #{companyName}, #{companyEmail}, #{companyTel}, #{companyLocation}, #{companyUsername}, #{companyPassword})
 	</insert>
@@ -51,6 +51,6 @@
 
 	<delete id="deleteById">
 		DELETE FROM company WHERE company_id=#{companyId}
-	</delete>
+	</delete> -->
 
 </mapper>
