+++ src/main/resources/mapper/intro.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="site.metacoding.miniproject.domain.intro.IntroDao">
 
-	<insert id="insert">
+	<!-- <insert id="insert">
 		INSERT INTO intro
 		(company_id, intro_coname, intro_birth, intro_task, intro_sal, intro_wellfare, intro_content, intro_location, job_id, intro_image_id)
 		VALUES(#{companyId}, #{introConame}, #{introBirth}, #{introTask}, #{introSal}, #{introWellfare}, #{introContent}, #{introLocation}, #{jobId}, #{introImageId});	
@@ -74,6 +74,6 @@
 		LEFT OUTER JOIN company c
 		ON i.company_id = c.company_id
 		WHERE i.intro_id = #{introId}
-	</select>
+	</select> -->
 
 </mapper>
