+++ src/main/resources/mapper/job.xml
@@ -2,9 +2,9 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="site.metacoding.miniproject.domain.job.JobDao">
 
-    <select id="findAll" resultType="site.metacoding.miniproject.domain.job.Job">
+    <!-- <select id="findAll" resultType="site.metacoding.miniproject.domain.job.Job">
         select *
         FROM job
-    </select>
+    </select> -->
 
 </mapper>
